Clear the clippy warnings this round introduced
Seven warnings in code added on this branch (type complexity, is_multiple_of,
a redundant cast and clone, and a collapsible-if that is deliberately two ifs
because the outer one is comptime). The remaining 44 warnings from
`cargo clippy --all-targets` are in code this round did not touch.

+++ examples/measure_reliability.rs
@@ -102,7 +102,7 @@ fn main() -> Result<()> {
         (Some(a), Some(b), Some(c), Some(d)) => println!(
             "reference memory: snapshot reserved {:+.2} MiB; carried cache {:.3} MiB; \
              reserved across {REPEATS} scored windows {:+.2} MiB (pool total {:.1} MiB)",
-            mib(b) as f64 - mib(a),
+            mib(b) - mib(a),
             mib(cache_bytes as u64),
             mib(d) - mib(c),
             mib(d),
